helpers/transform.py file was organized into multiple files. Notebooks were adjusted
